Fixes a bug that misplaces observations.
The second to largest bucket was previously updated only about half of
the time. Instead of using the correct bucket boundary, the mu value for
the final bucket was used to determine the boundary. As a consequence,
that second to last bucket only received about half of its items -- the
ones below mu.