Implemented the main distance algorithms.
The distance metrics look reasonable, but the histograms don't visually
have the same area. That's probably a bug somewhere... Also, the
buckets on the edge are a bit flaky. On the exponential distribution,
we have a bucket that spans a good distance to the left of 0... which
shouldn't happen. Even if that bucket does extend that direction, it
should be limited in how far it extends... huh.